Test 3.7 and 3.8, remove utils file
