+++ .github/workflows/main.yml
@@ -15,7 +15,7 @@ jobs:
       - name: Set up Python 3.8
         uses: actions/setup-python@v1
         with:
-          python-version: 3.8
+          python-version: [3.7, 3.8]
 
       - name: Install Poetry
         uses: dschep/install-poetry-action@v1.2
